Put Python code for tools in `gradbench.tools` (#308)
