+++ python/gradbench/gradbench/tools/pytorch/gmm.py
@@ -20,88 +20,68 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-# https://github.com/microsoft/ADBench/blob/38cb7931303a830c3700ca36ba9520868327ac87/src/python/modules/Tensorflow/TensorflowGMM.py
+# https://github.com/microsoft/ADBench/blob/38cb7931303a830c3700ca36ba9520868327ac87/src/python/modules/PyTorch/PyTorchGMM.py
 
 """
 Changes Made:
-- Added two functions to create a TensorflowGMM object and call calculate_objective and calculate_jacobian
+- Added two functions to create a PyTorchGMM object and call calculate_objective and calculate_jacobian
 - Added function to create GMMInput object
 - Import a decorator to convert input and outputs to necessary types
 """
 
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-import tensorflow as tf
+import numpy as np
+import torch
 
 from gradbench import wrap
 from gradbench.adbench.defs import Wishart
-from gradbench.adbench.gmm_data import GMMInput, GMMOutput
+from gradbench.adbench.gmm_data import GMMInput
 from gradbench.adbench.itest import ITest
-from gradbench.tensorflow.gmm_objective import gmm_objective
-from gradbench.tensorflow.utils import flatten, to_tf_tensor
+from gradbench.tools.pytorch.gmm_objective import gmm_objective
+from gradbench.tools.pytorch.utils import (
+    to_torch_tensor,
+    to_torch_tensors,
+    torch_jacobian,
+)
 
 
-class TensorflowGMM(ITest):
-    """Test class for GMM differentiation by Tensorflow using eager
-    execution."""
+class PyTorchGMM(ITest):
+    """Test class for GMM differentiation by PyTorch."""
 
     def prepare(self, input):
         """Prepares calculating. This function must be run before
         any others."""
 
-        self.alphas = to_tf_tensor(input.alphas)
-        self.means = to_tf_tensor(input.means)
-        self.icf = to_tf_tensor(input.icf)
+        self.inputs = to_torch_tensors(
+            (input.alphas, input.means, input.icf), grad_req=True
+        )
 
-        self.x = to_tf_tensor(input.x)
-        self.wishart_gamma = to_tf_tensor(input.wishart.gamma)
-        self.wishart_m = to_tf_tensor(input.wishart.m)
+        self.params = to_torch_tensors((input.x, input.wishart.gamma, input.wishart.m))
 
-        self.objective = tf.zeros(1)
-        self.gradient = tf.zeros(0)
+        self.objective = torch.zeros(1)
+        self.gradient = torch.empty(0)
 
     def output(self):
         """Returns calculation result."""
 
-        return GMMOutput(self.objective.numpy(), self.gradient.numpy())
+        return GMMOutput(self.objective.item(), self.gradient.numpy())
 
     def calculate_objective(self, times):
         """Calculates objective function many times."""
 
-        for _ in range(times):
-            self.objective = gmm_objective(
-                self.alphas,
-                self.means,
-                self.icf,
-                self.x,
-                self.wishart_gamma,
-                self.wishart_m,
-            )
+        for i in range(times):
+            self.objective = gmm_objective(*self.inputs, *self.params)
 
     def calculate_jacobian(self, times):
         """Calculates objective function jacobian many times."""
 
-        for _ in range(times):
-            with tf.GradientTape(persistent=True) as t:
-                t.watch(self.alphas)
-                t.watch(self.means)
-                t.watch(self.icf)
-
-                self.objective = gmm_objective(
-                    self.alphas,
-                    self.means,
-                    self.icf,
-                    self.x,
-                    self.wishart_gamma,
-                    self.wishart_m,
-                )
-
-            J = t.gradient(self.objective, (self.alphas, self.means, self.icf))
-            self.gradient = tf.concat([flatten(d) for d in J], 0)
+        for i in range(times):
+            self.objective, self.gradient = torch_jacobian(
+                gmm_objective, self.inputs, self.params
+            )
 
 
 def prepare_input(input):
-    py = TensorflowGMM()
+    py = PyTorchGMM()
     py.prepare(
         GMMInput(
             input["alpha"],
@@ -114,13 +94,19 @@ def prepare_input(input):
     return py
 
 
-@wrap.multiple_runs(pre=prepare_input, post=lambda x: x.numpy().tolist())
+@wrap.multiple_runs(
+    pre=prepare_input,
+    post=lambda x: x.tolist(),
+)
 def jacobian(py):
     py.calculate_jacobian(1)
     return py.gradient
 
 
-@wrap.multiple_runs(pre=prepare_input, post=lambda x: x.numpy().tolist())
+@wrap.multiple_runs(
+    pre=prepare_input,
+    post=lambda x: x.tolist(),
+)
 def objective(py):
     py.calculate_objective(1)
     return py.objective
