+++ tools/autograd/Dockerfile
@@ -7,5 +7,5 @@ RUN pip install autograd
 COPY python /home/gradbench/python
 ENV PYTHONPATH=/home/gradbench/python/gradbench
 
-ENTRYPOINT ["python", "/home/gradbench/python/gradbench/gradbench/autograd/run.py"]
+ENTRYPOINT ["python", "/home/gradbench/python/gradbench/gradbench/tools/autograd/run.py"]
 LABEL org.opencontainers.image.source=https://github.com/gradbench/gradbench
